Addition: Show price and timestamp when hovering over a graph.

+++ include/drawobjects/drawobject.h
@@ -68,8 +68,8 @@ typedef struct TextStyle {
 
 typedef struct Point {
 
-    unsigned short int x;
-    unsigned short int y;
+    float x;
+    float y;
 
 } Point;
